Time: 29 ms (86.19%), Space: 56.5 MB (75%) - LeetHub

+++ 1838-frequency-of-the-most-frequent-element/1838-frequency-of-the-most-frequent-element.java
@@ -3,21 +3,19 @@ class Solution {
         int n = nums.length;
         int i = 0;
         int j = 0;
-        int currSum = 0;
+        long currSum = 0;
         int max = 0;
         Arrays.sort(nums);
         while (j < n) {
-
             currSum += nums[j];
 
-            while ((long)nums[j] * (j - i + 1) - currSum > k) {
+            while ((long) nums[j] * (j - i + 1) - currSum > k) {
                 currSum -= nums[i];
                 i++;
             }
+
             max = Math.max(max, j - i + 1);
-
             j++;
-
         }
         return max;
     }
